Per feedback, use 'long long' for sequence number.

+++ db/sql.h
@@ -511,7 +511,7 @@ struct clnt_ddl_context {
 
 /* Client specific sql state */
 struct sqlclntstate {
-    int seqNo;                 /* Monotonically increasing sequence number
+    long long seqNo;           /* Monotonically increasing sequence number
                                 * assigned during dispatch.  This value is
                                 * not allowed to be zero.  Further, it must
                                 * be based on a value that is never reset
